Add MPI as a dependency and MpiCommunicator impl
It was trivial to adapt the euler example, which
is satisfying. There are no deadlocks or network
errors any more, but there's still seemingly
duplicate messages coming on on the executor.
That's clearly a bug in the executor logic,
and it should be easier to identify without
the complication of experimental comm's.