Replace Rust MPI crate with boilerplate
The cost of including the MPI Rust crate on
the build system is crazy. It pulls in a ton
of extra crates, adds a lot of bug-prone
intermediate build products, and needs access
to an llvm library, automake, and makeinfo.

The alternative seems dramatically more
acceptable if it will work: just build a
tiny C wrapper with an FFI and an unsafe
Rust module. It compiles fine but still
needs to be tested.